requirements.txt
Flask==2.2.2
Pillow==9.4.0
Pillow==10.3.0
torch==2.1.1
torchvision==0.16.0
